Avoid triggering graph guard cond after destroying it (#3)
The graph guard condition is triggered whenever the underlying DDS
topology changes using a listener set on readers for the DDS built-in
topics.  The graph guard condition was destroyed before ensuring the listeners
would no longer be invoked, and this would lead to trying to trigger the
graph guard condition after it had been destroyed.

With this commit, the built-in readers are deleted explicitly before
destroying the graph guard condition.