Delete dds domain in rmw_context_impl_t destructor
The gcdds destructur always runs upon termination of the process, also
when the process hasn't deleted the RMW objects first, and so possibly
before the rmw_context_impl_t destructor runs.  As a consequence the
discovery thread may never be stopped and run into an error from
dds_waitset_wait because the waitset has been deleted.

Deleting the domain in the destructor of rmw_context_impl_t avoids this
problem.

Signed-off-by: Erik Boasson <eb@ilities.com>