Suppress cppcheck unknownMacro
Fast forward to https://github.com/ament/ament_lint/pull/268

This suppresses a defect in `ament_cppcheck` - namely that if a macro is missing, it fails loudly. This is compounded by fact that `ament_cppcheck` does not include all dependent headers, so macros are likely to be missing in the first place.

Signed-off-by: Dan Rose <dan@digilabs.io>