Update Galactic to support zero-copy (#321)
* Update sample size when constructing a sertype

Signed-off-by: Sumanth Nirmal <sumanth.724@gmail.com>

* Fix method to check if the type is fixed

Signed-off-by: Sumanth Nirmal <sumanth.724@gmail.com>

* Fix Publish/Take serialized path with SHM.

Loan memory from iceoryx, copy the de-serialized message into loaned memory
and transfer the data using iceoryx. Take the data from loaned memory,
serialize the message based on the take path and eventually free the
loaned memory chunk.

Signed-off-by: Sumanth Nirmal <sumanth.724@gmail.com>

* Store typesupport with cdds_subscription

Signed-off-by: Sumanth Nirmal <sumanth.724@gmail.com>

* Perform on demand serialization

Serialization is done when required

Signed-off-by: Sumanth Nirmal <sumanth.724@gmail.com>

* Fix the Take loan path when the message is published not using SHM

The message ends up in take_loan, but the message is not published over
SHM. Allocate memory for the message, de-serialize the sample into the
memory and return the memory pointer. This memory will be freed in the
subsequent call to return_loaned_message

Signed-off-by: Sumanth Nirmal <sumanth.724@gmail.com>

* Access the sample only when the received sample has the valid data

Signed-off-by: Sumanth Nirmal <sumanth.724@gmail.com>

* Handle the regular take path when the sample is transferred using iceoryx

  Copy the Sample from iceoryx chunk to ros_message and release the
chunk back to iceoryx

Signed-off-by: Sumanth Nirmal <sumanth.724@gmail.com>

* Serialize the sample in rmw_print on demand

When trace category is enabled in logging of Cyclone, this will print the
serdata sample. If the sample is transferred using iceoryx, serialize the
sampled when required, which will be printed in the log

Signed-off-by: Sumanth Nirmal <sumanth.724@gmail.com>

* Add functions for allocating and freeing memory for sample from Cyclone

Signed-off-by: Sumanth Nirmal <sumanth.724@gmail.com>

* Fix implicit size conversion with iox_size

Signed-off-by: Sumanth Nirmal <sumanth.724@gmail.com>

* Fix the compilation with SHM disabled

Signed-off-by: Sumanth Nirmal <sumanth.724@gmail.com>

* Cleanup of unnecessary compile definition check DDS_HAS_SHM

Signed-off-by: Sumanth Nirmal <sumanth.724@gmail.com>

* Fix typo in the error message

Co-authored-by: eboasson <eb@ilities.com>

* Update is_type_self_contained

Signed-off-by: Sumanth Nirmal <sumanth.724@gmail.com>

* Fix the data path for receiving the data when data is transferred using SHM

Perform serialization in `to_sample` when required instead of copying
the iox_chunk directly to the ros message

Signed-off-by: Sumanth Nirmal <sumanth.724@gmail.com>

Co-authored-by: Sumanth Nirmal <sumanth.724@gmail.com>