Minor revamp of the CMakeLists.txt.
The most important bit in here is to set
CMAKE_CXX_STANDARD_REQUIRED to ON, so it builds in C++17
mode with clang.  Besides that, cleanup a comment, and
get rid of extra calls to target_link_libraries that
aren't required (we can just combine that all into one
call).

Signed-off-by: Chris Lalancette <clalancette@gmail.com>