Remove a bunch of unnecessary macros.
The main goal here is to sidestep differences between
uncrustify 0.72 and 0.78.  However, I realized that a
lot of these macros were either duplicating functionality
we already had, or were just as long as "inlining" the code
would be.  Just get rid of these macros.

In the one place that we can't do this, just mark the
macro as not to be indented by uncrustify.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>