New deserializer based on the used serializer
There are currently two serializers:

- one in TypeSupport_impl.hpp
- one in Serializer.cpp

The first one is the one originally used in the Cyclone RMW layer and walks the ROS
message definition directly. The second one was built to speed up serialization and
operates on a data structure derived from the ROS message definition.

For several years, the second serializer has been used. The deserializer never got
rewritten, leading to a very messy situation. This introduces a new deserializer written
to match the serializer actually being used.